Update the home page ui.
